Motion Identifier Problem
This file contains code for a type of motion identifier. It asks for an input of velocity and acceleration, converts it all into m/s and m/s² respectfully, then identifies what type of motion it is. The choices could be, "Uniform Motion," "Accelerated Motion," "Decelerated Motion," or "At Rest."